refactoring and code changes
